refactoring input formats
