+++ twister2/data/src/main/java/edu/iu/dsc/tws/data/api/InputFormat.java
@@ -11,7 +11,6 @@
 //  limitations under the License.
 package edu.iu.dsc.tws.data.api;
 
-import java.io.IOException;
 import java.io.Serializable;
 
 import edu.iu.dsc.tws.common.config.Config;
@@ -75,46 +74,4 @@ public interface InputFormat<OT, T extends InputSplit> extends InputSplitter<T>,
 
   @Override
   InputSplitAssigner getInputSplitAssigner(T[] inputSplits);
-
-  /**
-   * Opens a parallel instance of the input format to work on a split.
-   * <p>
-   * When this method is called, the input format it guaranteed to be configured.
-   *
-   * @param split The split to be opened.
-   * @throws IOException Thrown, if the spit could not be opened due to an I/O problem.
-   */
-  void open(T split) throws IOException;
-
-  /**
-   * Method used to check if the end of the input is reached.
-   * <p>
-   * When this method is called, the input format it guaranteed to be opened.
-   *
-   * @return True if the end is reached, otherwise false.
-   * @throws IOException Thrown, if an I/O error occurred.
-   */
-  boolean reachedEnd() throws IOException;
-
-  /**
-   * Reads the next record from the input.
-   * <p>
-   * When this method is called, the input format it guaranteed to be opened.
-   *
-   * @param reuse Object that may be reused.
-   * @return Read record.
-   * @throws IOException Thrown, if an I/O error occurred.
-   */
-  OT nextRecord(OT reuse) throws IOException;
-
-  /**
-   * Method that marks the end of the life-cycle of an input split. Should be used to
-   * close channels and streams and release resources. After this method returns without an error,
-   * the input is assumed to be correctly read.
-   * <p>
-   * When this method is called, the input format it guaranteed to be opened.
-   *
-   * @throws IOException Thrown, if the input could not be closed properly.
-   */
-  void close() throws IOException;
 }
