+++ twister2/resource-scheduler/src/java/edu/iu/dsc/tws/rsched/uploaders/hdfs/HdfsUploader.java
@@ -10,35 +10,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package edu.iu.dsc.tws.rsched.uploaders.hdfs;
-
-/*import java.io.File;
-import java.net.URI;
-import java.util.logging.Logger;
-
-import com.twitter.heron.common.basics.TypeUtils;
-
-
-import com.twitter.heron.spi.common.Config;
-import com.twitter.heron.spi.common.Context;
-
-
-import com.twitter.heron.spi.uploader.IUploader;
-import com.twitter.heron.spi.uploader.UploaderException;
-import com.twitter.heron.spi.utils.UploaderUtils;
-*/
 
 import java.io.File;
 import java.net.URI;
@@ -48,9 +20,6 @@ import java.util.logging.Logger;
 import edu.iu.dsc.tws.common.config.Config;
 import edu.iu.dsc.tws.rsched.exceptions.UploaderException;
 import edu.iu.dsc.tws.rsched.interfaces.IUploader;
-
-
-
 
 public class HdfsUploader implements IUploader {
   private static final Logger LOG = Logger.getLogger(HdfsUploader.class.getName());
@@ -71,9 +40,6 @@ public class HdfsUploader implements IUploader {
     this.controller = getHdfsController();
 
     this.destTopologyDirectoryURI = HdfsContext.hdfsTopologiesDirectoryURI(config);
-
-
-
   }
 
   // Utils method
